Ensure Bordering Settlement

+++ include/populate.h
@@ -15,6 +15,6 @@
  * @param data The buffer containing the data
  * @return The new population of the cell
  */
-int calc_cell_population(int target_cell, struct DataDims data_dims, struct GhostCols ghost_cols, unsigned short *data);
+unsigned short calc_cell_population(int target_cell, struct DataDims data_dims, struct GhostCols ghost_cols, unsigned short *data);
 
 #endif //MANIFEST_DESTINY_POPULATE_H
